requirements.txt
streamlit
crewai
openai
anthropic
pydantic
pdfplumber
plotly
python-dotenv